% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{country_name}
\alias{country_name}
\title{Convert Country Codes to Country and Patent Instrument Names}
\usage{
country_name(x, col)
}
\arguments{
\item{x}{a data frame}

\item{col}{column containing two letter country codes}
}
\value{
data frame
}
\description{
This takes a field containing a two letter country/patent instrument code and converts it to a country or instrument name based on data from the country code packge.
}
\examples{
\dontrun{res <- country_name(family, "family_country")}
}
