% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Literature or Patent Data from the Lens}
\usage{
import(path, type = NULL)
}
\arguments{
\item{path}{A path to the file to import}

\item{type}{The type of data to import (lit or patent), defaults to lit.}
}
\value{
data frame
}
\description{
Import csv files downloaded from the Lens COVID collection of scientific or patent literature. If patent data specify patent in the type. Otherwise
  leave as is.
}
\details{
This code imports literature or patent data and converts col names to lower, replaces spaces with underscores and removes punctuation and brackets.
Patent publication and applicationnumber from the lens include spaces between the country code, the number and the kind code as well as "/" in the identifier. This
can cause problems if linking to other databases. The original Lens versions are maintained with the prefix lens_ and simplified versions are created lacking spaces
and punctuation are created for publication and application numbers.
}
\examples{
{df <- import_lenslit(path = "/patentr/data/lenslit.csv")
df <- import_lenslit(path = "/patentr/data/lenslit.csv", type = "patent"}
}
